Move classes that implement data structures into separate header/implementation files.
