+++ python/boomer/boosting/cpp/data/triangular_number.h
@@ -1,0 +1,21 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "../../../common/cpp/data/types.h"
+
+
+namespace boosting {
+
+    /**
+     * Calculates and returns the n-th triangular number, i.e., the number of elements in a n times n triangle.
+     *
+     * @param n A scalar of type `uint32`, representing the order of the triangular number
+     * @return  A scalar of type `uint32`, representing the n-th triangular number
+     */
+    static inline uint32 triangularNumber(uint32 n) {
+        return (n * (n + 1)) / 2;
+    }
+
+}
