+++ python/boomer/boosting/cpp/data/vector_dense_label_wise.h
@@ -1,29 +1,23 @@
 /**
- * Provides classes that store the gradients and Hessians that have been calculated using a non-decomposable loss
- * function in matrices or vectors.
- *
  * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
  */
 #pragma once
 
-#include "../../common/cpp/indices.h"
+#include "../../../common/cpp/indices.h"
 
 
 namespace boosting {
 
     /**
-     * An one-dimensional vector that stores gradients and Hessians that have been calculated using a non-decomposable
-     * loss function in C-contiguous arrays. For each element in the vector a single gradient, but multiple Hessians are
-     * stored. In a vector that stores `n` gradients `(n * (n + 1)) / 2` Hessians are stored. The Hessians can be viewed
-     * as a symmetric Hessian matrix with `n` rows and columns.
+     * An one-dimensional vector that stores gradients and Hessians that have been calculated using a label-wise
+     * decomposable loss function in C-contiguous arrays. For each element in the vector a single gradient and Hessian
+     * is stored.
      */
-    class DenseExampleWiseStatisticVector {
+    class DenseLabelWiseStatisticVector {
 
         private:
 
-            uint32 numGradients_;
-
-            uint32 numHessians_;
+            uint32 numElements_;
 
             float64* gradients_;
 
@@ -32,23 +26,23 @@ namespace boosting {
         public:
 
             /**
-             * @param numGradients The number of gradients in the vector
+             * @param numElements The number of gradients and Hessians in the vector
              */
-            DenseExampleWiseStatisticVector(uint32 numGradients);
+            DenseLabelWiseStatisticVector(uint32 numElements);
 
             /**
-             * @param numGradients The number of gradients in the vector
-             * @param init         True, if all gradients and Hessians in the vector should be initialized with zero,
-             *                     false otherwise
+             * @param numElements   The number of gradients and Hessians in the vector
+             * @param init          True, if all gradients and Hessians in the vector should be initialized with zero,
+             *                      false otherwise
              */
-            DenseExampleWiseStatisticVector(uint32 numGradients, bool init);
+            DenseLabelWiseStatisticVector(uint32 numElements, bool init);
 
             /**
-             * @param vector A reference to an object of type `DenseExampleWiseStatisticVector` to be copied
+             * @param vector A reference to an object of type `DenseLabelWiseStatisticVector` to be copied
              */
-            DenseExampleWiseStatisticVector(const DenseExampleWiseStatisticVector& vector);
+            DenseLabelWiseStatisticVector(const DenseLabelWiseStatisticVector& vector);
 
-            ~DenseExampleWiseStatisticVector();
+            ~DenseLabelWiseStatisticVector();
 
             typedef float64* gradient_iterator;
 
@@ -115,18 +109,9 @@ namespace boosting {
             hessian_const_iterator hessians_cend() const;
 
             /**
-             * Returns a Hessian on the diagonal of the Hessian matrix that corresponds to the label at a specific
-             * position.
+             * Returns the number of gradients and Hessians in the vector.
              *
-             * @param row   The position
-             * @return      The Hessian that corresponds to the given position
-             */
-            float64 hessian_diagonal(uint32 pos) const;
-
-            /**
-             * Returns the number of gradients in the vector.
-             *
-             * @return The number of gradients
+             * @return The number of gradients and Hessians
              */
             uint32 getNumElements() const;
 
@@ -246,144 +231,6 @@ namespace boosting {
                             const PartialIndexVector& firstIndices, gradient_const_iterator secondGradientsBegin,
                             gradient_const_iterator secondGradientsEnd, hessian_const_iterator secondHessiansBegin,
                             hessian_const_iterator secondHessiansEnd);
-
-    };
-
-    /**
-     * A two-dimensional matrix that stores gradients and Hessians that have been calculated using a non-decomposable
-     * loss function in C-contiguous arrays. For each element at a certain row a single gradient, but multiple Hessians
-     * are stored. In a vector that stores `n` gradients `(n * (n + 1)) / 2` Hessians are stored. The Hessians can be
-     * viewed as a symmetric Hessian matrix with `n` rows and columns.
-     */
-    class DenseExampleWiseStatisticMatrix {
-
-        private:
-
-            uint32 numRows_;
-
-            uint32 numGradients_;
-
-            uint32 numHessians_;
-
-            float64* gradients_;
-
-            float64* hessians_;
-
-        public:
-
-            /**
-             * @param numRows       The number of rows in the matrix
-             * @param numGradients  The number of gradients per row
-             */
-            DenseExampleWiseStatisticMatrix(uint32 numRows, uint32 numGradients);
-
-            /**
-             * @param numRows       The number of rows in the matrix
-             * @param numGradients  The number of gradients per row
-             * @param init          True, if all gradients and Hessians in the matrix should be initialized with zero,
-             *                      false otherwise
-             */
-            DenseExampleWiseStatisticMatrix(uint32 numRows, uint32 numGradients, bool init);
-            
-            ~DenseExampleWiseStatisticMatrix();
-            
-            typedef float64* gradient_iterator;
-            
-            typedef const float64* gradient_const_iterator;
-            
-            typedef float64* hessian_iterator;
-            
-            typedef const float64* hessian_const_iterator;
-            
-            /**
-             * Returns a `gradient_iterator` to the beginning of the gradients at a specific row.
-             *
-             * @param row   The row
-             * @return      A `gradient_iterator` to the beginning of the given row 
-             */
-            gradient_iterator gradients_row_begin(uint32 row);
-            
-            /**
-             * Returns a `gradient_iterator` to the end of the gradients at a specific row.
-             *
-             * @param row   The row
-             * @return      A `gradient_iterator` to the end of the given row 
-             */
-            gradient_iterator gradients_row_end(uint32 row);
-            
-            /**
-             * Returns a `gradient_const_iterator` to the beginning of the gradients at a specific row.
-             *
-             * @param row   The row
-             * @return      A `gradient_const_iterator` to the beginning of the given row 
-             */
-            gradient_const_iterator gradients_row_cbegin(uint32 row) const;
-            
-            /**
-             * Returns a `gradient_const_iterator` to the end of the gradients at a specific row.
-             *
-             * @param row   The row
-             * @return      A `gradient_const_iterator` to the end of the given row 
-             */
-            gradient_const_iterator gradients_row_cend(uint32 row) const;
-
-            /**
-             * Returns a `hessian_iterator` to the beginning of the Hessians at a specific row.
-             *
-             * @param row   The row
-             * @return      A `hessian_iterator` to the beginning of the given row 
-             */
-            hessian_iterator hessians_row_begin(uint32 row);
-            
-            /**
-             * Returns a `hessian_iterator` to the end of the Hessians at a specific row.
-             *
-             * @param row   The row
-             * @return      A `hessian_iterator` to the end of the given row 
-             */
-            hessian_iterator hessians_row_end(uint32 row);
-            
-            /**
-             * Returns a `hessian_const_iterator` to the beginning of the Hessians at a specific row.
-             *
-             * @param row   The row
-             * @return      A `hessian_const_iterator` to the beginning of the given row 
-             */
-            hessian_const_iterator hessians_row_cbegin(uint32 row) const;
-            
-            /**
-             * Returns a `hessian_const_iterator` to the end of the Hessians at a specific row.
-             *
-             * @param row   The row
-             * @return      A `hessian_const_iterator` to the end of the given row 
-             */
-            hessian_const_iterator hessians_row_cend(uint32 row) const;
-
-            /**
-             * Returns the number of rows in the matrix.
-             *
-             * @return The number of rows
-             */
-            uint32 getNumRows() const;
-
-            /**
-             * Returns the number of gradients per row.
-             *
-             * @return The number of gradients
-             */
-            uint32 getNumCols() const;
-
-            /**
-             * Adds all gradients and Hessians in a vector to a specific row of this matrix.
-             *
-             * @param row               The row
-             * @param gradientsBegin    A `gradient_const_iterator` to the beginning of the gradients in the vector
-             * @param gradientsEnd      A `gradient_const_iterator` to the end of the gradients in the vector
-             * @param hessiansBegin     A `hessian_const_iterator` to the beginning of the Hessians in the vector
-             * @param hessiansEnd       A `hessian_const_iterator` to the end of the Hessians in the vector
-             */
-            void addToRow(uint32 row, gradient_const_iterator gradientsBegin, gradient_const_iterator gradientsEnd,
-                          hessian_const_iterator hessiansBegin, hessian_const_iterator hessiansEnd);
 
     };
