+++ python/boomer/boosting/cpp/losses_label_wise.h
@@ -6,8 +6,8 @@
 #pragma once
 
 #include "../../common/cpp/input_data.h"
-#include "data.h"
-#include "data_label_wise.h"
+#include "data/matrix_dense_numeric.h"
+#include "data/matrix_dense_label_wise.h"
 
 
 namespace boosting {
