+++ python/boomer/boosting/cpp/rule_evaluation_example_wise.cpp
@@ -1,5 +1,4 @@
 #include "rule_evaluation_example_wise.h"
-#include "data_example_wise.h"
 #include "linalg.h"
 
 using namespace boosting;
