+++ python/boomer/boosting/cpp/statistics_label_wise.cpp
@@ -1,6 +1,7 @@
 #include "statistics_label_wise.h"
-#include "data.h"
-#include "data_label_wise.h"
+#include "data/matrix_dense_numeric.h"
+#include "data/matrix_dense_label_wise.h"
+#include "data/vector_dense_label_wise.h"
 #include <memory>
 
 using namespace boosting;
