+++ python/boomer/common/cpp/data/indexed_value.h
@@ -1,0 +1,16 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "types.h"
+
+
+/**
+ * A tuple that consists of an index and a value.
+ */
+template<class T>
+struct IndexedValue {
+    uint32 index;
+    T value;
+};
