+++ python/boomer/common/cpp/data/vector_dok_binary.h
@@ -1,0 +1,57 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "types.h"
+#include <unordered_set>
+
+
+/**
+ * An one-dimensional sparse vector that stores binary data using the dictionary of keys (DOK) format.
+ */
+class BinaryDokVector {
+
+    private:
+
+        std::unordered_set<uint32> data_;
+
+    public:
+
+        typedef std::unordered_set<uint32>::const_iterator index_const_iterator;
+
+        /**
+         * Returns an `index_const_iterator` to the beginning of the indices.
+         *
+         * @return An `index_const_iterator` to the beginning
+         */
+        index_const_iterator indices_cbegin() const;
+
+        /**
+         * Returns an `index_const_iterator` to the end of the indices.
+         *
+         * @return An `index_const_iterator` to the end
+         */
+        index_const_iterator indices_cend() const;
+
+        /**
+         * Returns the value of the element at a specific position.
+         *
+         * @param pos   The position of the element. Must be in [0, getNumElements())
+         * @return      The value of the given element
+         */
+        bool getValue(uint32 pos) const;
+
+        /**
+         * Sets a non-zero value to the element at a specific position.
+         *
+         * @param pos The position of the element. Must be in [0, getNumElements())
+         */
+        void setValue(uint32 pos);
+
+        /**
+         * Sets the values of all elements to zero.
+         */
+        void setAllToZero();
+
+};
