+++ python/boomer/common/cpp/data/vector_sparse_array.h
@@ -1,0 +1,90 @@
+/**
+ * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
+ */
+#pragma once
+
+#include "types.h"
+#include "indexed_value.h"
+
+
+/**
+ * An one-dimensional sparse vector that stores a fixed number of elements, consisting of an index and a value, in a
+ * C-contiguous array.
+ *
+ * @tparam T The type of the data that is stored in the vector
+ */
+template<class T>
+class SparseArrayVector {
+
+    public:
+
+        typedef IndexedValue<T> Entry;
+
+
+    private:
+
+        Entry* array_;
+
+        uint32 numElements_;
+
+    public:
+
+        /**
+         * @param numElements The number of elements in the vector
+         */
+        SparseArrayVector(uint32 numElements);
+
+        ~SparseArrayVector();
+
+        typedef Entry* iterator;
+
+        typedef const Entry* const_iterator;
+
+        /**
+         * Returns an `iterator` to the beginning of the vector.
+         *
+         * @return An `iterator` to the beginning
+         */
+        iterator begin();
+
+        /**
+         * Returns an `iterator` to the end of the vector.
+         *
+         * @return An `iterator` to the end
+         */
+        iterator end();
+
+        /**
+         * Returns a `const_iterator` to the beginning of the vector.
+         *
+         * @return A `const_iterator` to the beginning
+         */
+        const_iterator cbegin() const;
+
+        /**
+         * Returns a `const_iterator` to the end of the vector.
+         *
+         * @return A `const_iterator` to the end
+         */
+        const_iterator cend() const;
+
+        /**
+         * Returns the number of elements in the vector.
+         *
+         * @return The number of elements in the vector
+         */
+        uint32 getNumElements() const;
+
+        /**
+         * Sets the number of elements in the vector.
+         *
+         * @param numElements The number of elements to be set
+         */
+        void setNumElements(uint32 numElements);
+
+        /**
+         * Sorts the elements in the vector in ascending order based on their values.
+         */
+        void sortByValues();
+
+};
