+++ python/boomer/common/cpp/indices.h
@@ -6,7 +6,7 @@
  */
 #pragma once
 
-#include "data.h"
+#include "data/vector_dense.h"
 #include <memory>
 
 // Forward declarations
