+++ python/boomer/common/cpp/random.h
@@ -5,7 +5,7 @@
  */
 #pragma once
 
-#include "types.h"
+#include "data/types.h"
 
 
 /**
