+++ python/boomer/common/cpp/rules.h
@@ -5,7 +5,7 @@
  */
 #pragma once
 
-#include "types.h"
+#include "data/types.h"
 #include <list>
 #include <array>
