+++ python/boomer/seco/cpp/confusion_matrices.h
@@ -4,7 +4,7 @@
  * @author Jakob Steeg (jakob.steeg@gmail.com)
  * @author Michael Rapp (mrapp@ke.tu-darmstadt.de)
  */
-#include "../../common/cpp/types.h"
+#include "../../common/cpp/data/types.h"
 
 
 namespace seco {
