+++ python/boomer/seco/cpp/lift_functions.h
@@ -7,7 +7,7 @@
  */
 #pragma once
 
-#include "../../common/cpp/types.h"
+#include "../../common/cpp/data/types.h"
 
 
 namespace seco {
